Венести функции в отдельный файл
