+++ getting_schedule/data_conversion.py
@@ -1,10 +1,14 @@
 from datetime import datetime, date
 import pytz
 
+from functions import schedule_tools
+
 TIME_ZONE = pytz.timezone('Asia/Irkutsk')
 
 # Режим отладки (если включен, то не определяем текущее время - позволяет использовать старое расписание).
-DEBUG = True
+DEBUG = False
+
+DAYS = schedule_tools.DAYS
 
 
 def convert_institutes(pg_institutes: list) -> list:
@@ -60,51 +64,6 @@ def convert_teachers(pg_teachers: list) -> list:
     return mongo_teachers
 
 
-DAYS = {
-    1: 'понедельник',
-    2: 'вторник',
-    3: 'среда',
-    4: 'четверг',
-    5: 'пятница',
-    6: 'суббота',
-    7: 'воскресенье'
-}
-
-
-def getting_week_and_day_of_week(pg_lesson: dict) -> tuple:
-    """Определение четности недели и дня недели"""
-
-    if pg_lesson['everyweek'] == 2:
-        week = 'all'
-        day = DAYS[pg_lesson['day']]
-    else:
-        if pg_lesson['day'] <= 7:
-            week = 'even'
-            day = DAYS[pg_lesson['day']]
-        else:
-            week = 'odd'
-            day = DAYS[pg_lesson['day'] - 7]
-
-    return week, day
-
-
-def is_there_dict_with_value_in_list(input_list_with_dict: list, value: str) -> bool:
-    if not input_list_with_dict:
-        return False
-
-    for dict_item in input_list_with_dict:
-        if value in dict_item.values():
-            return True
-    return False
-
-
-def get_dict_key(d, value):
-    """Получение ключа по значеню словаря"""
-    for k, v in d.items():
-        if v == value:
-            return k
-
-
 def convert_schedule(pg_schedule: list) -> list:
     """Преобразование формата расписания"""
 
@@ -126,21 +85,10 @@ def convert_schedule(pg_schedule: list) -> list:
         # Проверяем, что расписание действует
         if date_now <= item['dend']:
 
-            week, day = getting_week_and_day_of_week(item)
+            week, day = schedule_tools.getting_week_and_day_of_week(item)
 
             # Определяем вид пары и подгруппу.
-            if item['nt'] == 1:
-                info = '( Лекция )'
-            elif item['nt'] == 2:
-                if item['ngroup']:
-                    info = f'( Практ. подгруппа {item["ngroup"]} )'
-                else:
-                    info = '( Практ. )'
-            else:
-                if item['ngroup']:
-                    info = f'( Лаб. раб. подгруппа {item["ngroup"]} )'
-                else:
-                    info = f'( Лаб. раб. )'
+            info = schedule_tools.forming_info_data(nt=item['nt'], ngroup=item["ngroup"])
 
             lesson = {
                 'time': item['begtime'],
@@ -152,7 +100,7 @@ def convert_schedule(pg_schedule: list) -> list:
             }
 
             # Смотрим, создал ли уже нужный день в расписании.
-            if not is_there_dict_with_value_in_list(schedule, day):
+            if not schedule_tools.is_there_dict_with_value_in_list(schedule, day):
                 schedule.append(
                     {
                         'day': day,
@@ -178,15 +126,11 @@ def convert_schedule(pg_schedule: list) -> list:
             # Проверяем, что расписание не пустое
             if schedule:
                 # Сортируем пары в дне по времени и подгруппе
-                for sch in schedule:
-                    # Сортируем подгруппы
-                    sch['lessons'] = sorted(sch['lessons'], key=lambda x: x['info'])
-                    # Сортируем по времени
-                    sch['lessons'] = sorted(sch['lessons'], key=lambda x: int(x['time'].replace(':', '')))
+                schedule_tools.sorting_lessons_in_a_day_by_time_and_ngroup(schedule=schedule)
 
                 all_schedule.append({
                     'group': current_group,
-                    'schedule': sorted(schedule, key=lambda x: get_dict_key(DAYS, x['day']))
+                    'schedule': sorted(schedule, key=lambda x: schedule_tools.get_dict_key(DAYS, x['day']))
                 })
 
                 # Обнуляем расписание для слудующей группы
@@ -216,21 +160,10 @@ def convert_teachers_schedule(pg_schedule: list) -> list:
         # Проверяем, что расписание действует
         if date_now <= item['dend']:
 
-            week, day = getting_week_and_day_of_week(item)
+            week, day = schedule_tools.getting_week_and_day_of_week(item)
 
             # Определяем вид пары и подгруппу.
-            if item['nt'] == 1:
-                info = '( Лекция )'
-            elif item['nt'] == 2:
-                if item['ngroup']:
-                    info = f'( Практ. подгруппа {item["ngroup"]} )'
-                else:
-                    info = '( Практ. )'
-            else:
-                if item['ngroup']:
-                    info = f'( Лаб. раб. подгруппа {item["ngroup"]} )'
-                else:
-                    info = f'( Лаб. раб. )'
+            info = schedule_tools.forming_info_data(nt=item['nt'], ngroup=item["ngroup"])
 
             lesson = {
                 'time': item['begtime'],
@@ -242,7 +175,7 @@ def convert_teachers_schedule(pg_schedule: list) -> list:
             }
 
             # Смотрим, создал ли уже нужный день в расписании.
-            if not is_there_dict_with_value_in_list(schedule, day):
+            if not schedule_tools.is_there_dict_with_value_in_list(schedule, day):
                 schedule.append(
                     {
                         'day': day,
@@ -279,18 +212,13 @@ def convert_teachers_schedule(pg_schedule: list) -> list:
         if current_prep_id != next_prep_id or item_index == len(pg_schedule) - 1:
             # Проверяем, что расписание не пустое
             if schedule:
-                # Сортируем пары в дне по времени и подгруппе
-                for sch in schedule:
-                    # Сортируем подгруппы
-                    sch['lessons'] = sorted(sch['lessons'], key=lambda x: x['info'])
-                    # Сортируем по времени
-                    sch['lessons'] = sorted(sch['lessons'], key=lambda x: int(x['time'].replace(':', '')))
+                schedule_tools.sorting_lessons_in_a_day_by_time_and_ngroup(schedule=schedule)
 
                 all_schedule.append({
                     'prep': item['preps'].strip(),
                     'prep_short_name': item['prep_short_name'].strip(),
                     'pg_id': current_prep_id,
-                    'schedule': sorted(schedule, key=lambda x: get_dict_key(DAYS, x['day']))
+                    'schedule': sorted(schedule, key=lambda x: schedule_tools.get_dict_key(DAYS, x['day']))
                 })
 
                 # Обнуляем расписание для слудующего преподавателя
